[Grupo 24]  altera o price dos products na model de FLOAT para DECIMAL

+++ back-end/src/database/models/products.js
@@ -1,7 +1,7 @@
 const products = (sequelize, DataTypes) => {
   const product = sequelize.define('product', {
     name: DataTypes.STRING,
-    price: DataTypes.FLOAT,
+    price: DataTypes.DECIMAL,
     url_image: DataTypes.STRING,
   },
   {tableName: 'products'});
